getBaseUrl updated to handle production env
