+++ site/src/app/api/room/createRoom.ts
@@ -1,5 +1,4 @@
 import { getBaseUrl } from '@/lib/server/getBaseUrl';
-// import { scheduleRoomDeletion } from './deleteRoom'; // Import the scheduled deletion function
 
 type room = {
 	room_id: number;
@@ -13,24 +12,13 @@ export async function createRoom() {
 		method: 'POST',
 	});
 
-	// On failed request, return null middleware for redirect
+	// On failed request, retry, don't return
 	if (res.status == 500) {
 		return null;
 	}
 
-	// Set the created_at timestamp using Prisma
-	// await prisma.rooms.update({
-	// 	where: {
-	// 		room_id: room.room_id,
-	// 	},
-	// 	data: {
-	// 		created_at: new Date(),
-	// 	} as any, // Use a type assertion here to handle the type mismatch
-	// });
-
 	// Schedule a deletion task for 24 hours in the future
 	const room = (await res.json()) as room;
-	// scheduleRoomDeletion(room.room_id, 24 * 60 * 60 * 1000); // 24 hours in milliseconds
 
 	return room.room_id;
 }
