+++ site/src/app/api/room/route.ts
@@ -17,3 +17,22 @@ export async function POST(req: Request) {
 		return new NextResponse('Error creating room', { status: 500 });
 	}
 }
+
+export async function DELETE(req: Request) {
+	try {
+		const res = await prisma.rooms.deleteMany({
+			where: {
+				created_at: {
+					// 24 * 60 * 30 * 1000 = 86400000 miliseconds = 24 hours
+					// 30,000 = 30 seconds for testing
+					lte: new Date(Date.now() - 30000),
+				},
+			},
+		});
+		console.log('successfully deleted expired rooms!');
+		return NextResponse.json(res);
+	} catch (error) {
+		console.error(`Error deleting expired rooms`, error);
+		return new NextResponse('Error deleting expired rooms', { status: 500 });
+	}
+}
