+++ site/src/lib/server/getBaseUrl.ts
@@ -1,5 +1,7 @@
-export function getBaseUrl() {
-	// Serverside function to get the base url & port
-	const baseUrl = 'http://localhost:3000';
-	return baseUrl;
+export function getBaseUrl(): string {
+	const host =
+		process.env.NODE_ENV === 'production'
+			? process.env.NEXT_PUBLIC_SITE_URL!
+			: `http://localhost:${process.env.PORT ?? 3000}`;
+	return host;
 }
